refactor: marketPriceApi and commodityUnitsApi endpoints
